Extend stop control to the clarify (Ask) footer
