+++ packages/ballerina-visualizer/src/views/AIPanel/components/AIChat/Footer/StopControl.tsx
@@ -1,0 +1,29 @@
+/**
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import React from "react";
+import { FooterIconBtn } from "./styles";
+
+// AIChatInput normally owns Escape-to-stop, but it's unmounted while an approval/clarify footer shows.
+const StopControl: React.FC<{ onStop: () => void }> = ({ onStop }) => (
+    <FooterIconBtn onClick={onStop} title="Stop (Escape)">
+        <span className="codicon codicon-stop-circle" />
+    </FooterIconBtn>
+);
+
+export default StopControl;
